+++ module/actes-generique/action/TedetisVerifReponsePref.class.php
@@ -10,9 +10,9 @@ class TedetisVerifReponsePref extends ActionExecutor {
 		if (!$tdT){
 			throw new Exception("Aucun Tdt disponible");
 		}
-		$tedetis_transaction_id = $this->getDonneesFormulaire()->get('tedetis_transaction_id');
-			
-		$all_response = $tdT->getListReponsePrefecture($tedetis_transaction_id);
+		$tedetis_transaction_id = $this->getDonneesFormulaire()->get('related_transaction_id');
+        /** @var S2low $tdT */
+        $all_response = $tdT->getListReponsePrefecture($tedetis_transaction_id);
 		
 		if (!$all_response)  {
 			$this->setLastMessage("Aucune réponse disponible");